backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
google-genai==1.25.0
google-auth==2.40.3 